Overview: Completed the testing of data storage functions on the topic of contacts and recruiters.
Fleshed out more of the mechanics and this just needs to be ported over to the server on the case of uploading *.csv files and then merging said *.csv files with existing contacts.

Process should follow as such: Recruiter information is typed in or chosen from the database. Following this, the user has the choice to manually enter individual contact information, upload entire *.csv files, or create a new recruiter elsewhere, enter that information, and enter things from there.